connection error message implemented

+++ app/Repositories/AbstractRepository.php
@@ -2,9 +2,12 @@
 
 namespace App\Repositories;
 
+use Illuminate\Support\Facades\DB;
+
 abstract class AbstractRepository
 {
     protected $model;
+    protected $pivot = '';
 
     /**
      * Assigns the child's model class to the template attribute.
@@ -25,14 +28,15 @@ abstract class AbstractRepository
     }
 
     /**
-     * Crate new register in database
+     *  Returns the model class used by the child class
      *
-     * @return 
+     * @return \App\Models
      */
-    public function create($data)
+    protected function resolvePivot()
     {
-        return $this->model->create($data);
+        return app($this->model->pivot);
     }
+
 
     /**
      * Return all data
@@ -41,7 +45,31 @@ abstract class AbstractRepository
      */
     public function all()
     {
+        if (!$this->testConnection()) {
+            return false;
+        }
+
+        if($this->pivot){
+            return $this->model->with($this->pivot)->get();
+        }
+
         return $this->model->all();
+
+    }
+
+    /**
+     * Crate new register in database
+     *
+     * @return Collection
+     */
+    public function create($data)
+    {
+        if ($this->testConnection()) {
+            return $this->model->create($data);
+        }
+
+        return false;
+
     }
 
     /**
@@ -51,7 +79,46 @@ abstract class AbstractRepository
      */
     public function find(int $id)
     {
+        if (!$this->testConnection()) {
+             return false;
+        }
+        if($this->pivot){
+
+            $search = $this->model->where('id',$id)->with($this->pivot)->get();
+
+            return count($search) <= 0 ? null : $search;
+        }
         return $this->model->find($id);
+
+
+    }
+
+    /**
+     * update a register in database
+     *
+     * @return Collection
+     */
+    public function update($request, $id)
+    {
+
+        $instanceModel = $this->find($id);
+// dd($instanceModel);
+        if (!$instanceModel) {
+            return false;
+        }
+        if ($instanceModel === null) {
+            return null;
+        }
+
+        if ($this->testConnection()) {
+
+            $instanceModel->fill($request->all());
+            $instanceModel->save();
+
+            return $instanceModel;
+
+        }
+        return false;
     }
 
     /**
@@ -61,6 +128,23 @@ abstract class AbstractRepository
      */
     public function destroy(int $id)
     {
-        return $this->model->destroy($id);
+        if ($this->testConnection()) {
+            return $this->model->destroy($id);
+        }
+        return false;
+
+    }
+
+    public function testConnection()
+    {
+        try {
+            if (DB::connection()->getPdo()) {
+                return true;
+            }
+
+        } catch (\PDOException $e) {
+            return false;
+
+        }
     }
 }
